+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.13.0 (xx.x.xxxx)
+
+- first take on removing the global route state => instead of global `swagger` atom, there is one defined `+routes+` var per namespace
+  - requires a `compojure.api.core/with-routes` on api root to generate and hold the `+routes+` (automatically bundled with `defapi`)
+
 ## 0.12.0 (17.6.2014)
 
 - **possibly breaking change**: `middlewares` macro and `:middlewares` restructuring now use thread-first to apply middlewares